Revert "Revert "Revert "Revert "Redid the whole Collision Detection system and converted all destination rectangles to use doubles instead of ints""""
This reverts commit 7de4e96af1fc7e0a7f8532414a73c17209134500.

+++ WindowsGame1/WindowsGame1/MapAsset.cs
@@ -21,7 +21,7 @@ namespace Spaceman
 
 		public Spaceman offsetSpaceMan(Spaceman s, int xOffset, int yOffset)
 		{
-			Spaceman result = new Spaceman(s.body, s.head,new Vector2(s.destRect.X, s.destRect.Y),s.numFrames, s.frameNum, s.mirrorX);
+			Spaceman result = new Spaceman(s.body, s.head,new Vector2((float)s.destRect.X, (float)s.destRect.Y),s.numFrames, s.frameNum, s.mirrorX);
 			result.destRect.X += xOffset;
 			result.destRect.Y += yOffset;
 			return result;
